test(grid): Move keyboard navigation tests in one file #2975
