+++ projects/igniteui-angular/src/lib/grids/grid/grid-selection.spec.ts
@@ -10,20 +10,10 @@ import { IgxStringFilteringOperand } from '../../../public_api';
 import { UIInteractions, wait } from '../../test-utils/ui-interactions.spec';
 import { IgxNumberFilteringOperand } from '../../data-operations/filtering-condition';
 import { configureTestSuite } from '../../test-utils/configure-suite';
+import { ScrollsComponent, GridWithPrimaryKeyComponent, SelectionComponent } from '../../test-utils/grid-samples.spec';
+import { SampleTestData } from '../../test-utils/sample-test-data.spec';
 
 const selectedCellClass = '.igx-grid__td--selected';
-let data = [
-    {ID: 1, Name: 'Casey Houston', JobTitle: 'Vice President', HireDate: '2017-06-19T11:43:07.714Z'},
-    {ID: 2, Name: 'Gilberto Todd', JobTitle: 'Director', HireDate: '2015-12-18T11:23:17.714Z'},
-    {ID: 3, Name: 'Tanya Bennett', JobTitle: 'Director', HireDate: '2005-11-18T11:23:17.714Z'},
-    {ID: 4, Name: 'Jack Simon', JobTitle: 'Software Developer', HireDate: '2008-12-18T11:23:17.714Z'},
-    {ID: 5, Name: 'Celia Martinez', JobTitle: 'Senior Software Developer', HireDate: '2007-12-19T11:23:17.714Z'},
-    {ID: 6, Name: 'Erma Walsh', JobTitle: 'CEO', HireDate: '2016-12-18T11:23:17.714Z'},
-    {ID: 7, Name: 'Debra Morton', JobTitle: 'Associate Software Developer', HireDate: '2005-11-19T11:23:17.714Z'},
-    {ID: 8, Name: 'Erika Wells', JobTitle: 'Software Development Team Lead', HireDate: '2005-10-14T11:23:17.714Z'},
-    {ID: 9, Name: 'Leslie Hansen', JobTitle: 'Associate Software Developer', HireDate: '2013-10-10T11:23:17.714Z'},
-    {ID: 10, Name: 'Eduardo Ramirez', JobTitle: 'Manager', HireDate: '2011-11-28T11:23:17.714Z'}
-];
 
 describe('IgxGrid - Row Selection', () => {
     configureTestSuite();
@@ -33,9 +23,9 @@ describe('IgxGrid - Row Selection', () => {
             declarations: [
                 GridWithPrimaryKeyComponent,
                 GridWithPagingAndSelectionComponent,
-                GridWithSelectionComponent,
+                SelectionComponent,
                 GridWithSelectionFilteringComponent,
-                GridWithScrollsComponent,
+                ScrollsComponent,
                 GridSummaryComponent,
                 GridCancelableComponent
             ],
@@ -45,24 +35,12 @@ describe('IgxGrid - Row Selection', () => {
             ]
         })
             .compileComponents();
-        data = [
-            {ID: 1, Name: 'Casey Houston', JobTitle: 'Vice President', HireDate: '2017-06-19T11:43:07.714Z'},
-            {ID: 2, Name: 'Gilberto Todd', JobTitle: 'Director', HireDate: '2015-12-18T11:23:17.714Z'},
-            {ID: 3, Name: 'Tanya Bennett', JobTitle: 'Software Developer', HireDate: '2005-11-18T11:23:17.714Z'},
-            {ID: 4, Name: 'Jack Simon', JobTitle: 'Senior Software Developer', HireDate: '2008-12-18T11:23:17.714Z'},
-            {ID: 5, Name: 'Celia Martinez', JobTitle: 'CEO', HireDate: '2007-12-19T11:23:17.714Z'},
-            {ID: 6, Name: 'Erma Walsh', JobTitle: 'CEO', HireDate: '2016-12-18T11:23:17.714Z'},
-            {ID: 7, Name: 'Debra Morton', JobTitle: 'Associate Software Developer', HireDate: '2005-11-19T11:23:17.714Z'},
-            {ID: 8, Name: 'Erika Wells', JobTitle: 'Software Development Team Lead', HireDate: '2005-10-14T11:23:17.714Z'},
-            {ID: 9, Name: 'Leslie Hansen', JobTitle: 'Associate Software Developer', HireDate: '2013-10-10T11:23:17.714Z'},
-            {ID: 10, Name: 'Eduardo Ramirez', JobTitle: 'Manager', HireDate: '2011-11-28T11:23:17.714Z'}
-        ];
     }));
 
     it('Should be able to select row through primaryKey and index', () => {
         const fix = TestBed.createComponent(GridWithPrimaryKeyComponent);
         fix.detectChanges();
-        const grid = fix.componentInstance.gridSelection1;
+        const grid = fix.componentInstance.grid;
 
         expect(grid.primaryKey).toBeTruthy();
         expect(grid.rowList.length).toEqual(10, 'All 10 rows should initialized');
@@ -73,7 +51,7 @@ describe('IgxGrid - Row Selection', () => {
     it('Should be able to update a cell in a row through primaryKey', () => {
         const fix = TestBed.createComponent(GridWithPrimaryKeyComponent);
         fix.detectChanges();
-        const grid = fix.componentInstance.gridSelection1;
+        const grid = fix.componentInstance.grid;
         expect(grid.primaryKey).toBeTruthy();
         expect(grid.rowList.length).toEqual(10, 'All 10 rows should initialized');
         expect(grid.getRowByKey(2).rowData['JobTitle']).toMatch('Director');
@@ -85,7 +63,7 @@ describe('IgxGrid - Row Selection', () => {
     it('Should be able to update row through primaryKey', () => {
         const fix = TestBed.createComponent(GridWithPrimaryKeyComponent);
         fix.detectChanges();
-        const grid = fix.componentInstance.gridSelection1;
+        const grid = fix.componentInstance.grid;
         spyOn(grid.cdr, 'markForCheck').and.callThrough();
         expect(grid.primaryKey).toBeTruthy();
         expect(grid.rowList.length).toEqual(10, 'All 10 rows should initialized');
@@ -100,7 +78,7 @@ describe('IgxGrid - Row Selection', () => {
     it('Should be able to delete a row through primaryKey', () => {
         const fix = TestBed.createComponent(GridWithPrimaryKeyComponent);
         fix.detectChanges();
-        const grid = fix.componentInstance.gridSelection1;
+        const grid = fix.componentInstance.grid;
         expect(grid.primaryKey).toBeTruthy();
         expect(grid.rowList.length).toEqual(10, 'All 10 rows should initialized');
         expect(grid.getRowByKey(2)).toBeDefined();
@@ -113,7 +91,7 @@ describe('IgxGrid - Row Selection', () => {
     it('Should handle update by not overwriting the value in the data column specified as primaryKey', () => {
         const fix = TestBed.createComponent(GridWithPrimaryKeyComponent);
         fix.detectChanges();
-        const grid = fix.componentInstance.gridSelection1;
+        const grid = fix.componentInstance.grid;
         expect(grid.primaryKey).toBeTruthy();
         expect(grid.rowList.length).toEqual(10, 'All 10 rows should initialized');
         expect(grid.getRowByKey(2)).toBeDefined();
@@ -123,8 +101,6 @@ describe('IgxGrid - Row Selection', () => {
         expect(grid.getRowByIndex(1)).toBeDefined();
         expect(grid.getRowByIndex(1).rowData[grid.primaryKey]).toEqual(7);
     });
-
-  
 
     it('Should persist through paging', (async () => {
         const fix = TestBed.createComponent(GridWithPagingAndSelectionComponent);
@@ -234,9 +210,9 @@ describe('IgxGrid - Row Selection', () => {
 
     it('Should persist through scrolling', (async () => {
         let selectedCell;
-        const fix = TestBed.createComponent(GridWithSelectionComponent);
+        const fix = TestBed.createComponent(SelectionComponent);
         fix.detectChanges();
-        const grid = fix.componentInstance.gridSelection3;
+        const grid = fix.componentInstance.grid;
         const gridElement: HTMLElement = fix.nativeElement.querySelector('.igx-grid');
         const selectedRow = grid.getRowByIndex(0);
         expect(selectedRow).toBeDefined();
@@ -306,10 +282,10 @@ describe('IgxGrid - Row Selection', () => {
     }));
 
     it('Should handle the deleteion on a selected row propertly', (async () => {
-        const fix = TestBed.createComponent(GridWithSelectionComponent);
+        const fix = TestBed.createComponent(SelectionComponent);
         fix.detectChanges();
 
-        const grid = fix.componentInstance.gridSelection3;
+        const grid = fix.componentInstance.grid;
         const headerRow: HTMLElement = fix.nativeElement.querySelector('.igx-grid__thead');
         const firstRow = grid.getRowByKey('0_0');
         const firstRowCheckbox: HTMLInputElement = firstRow.nativeElement.querySelector('.igx-checkbox__input');
@@ -415,9 +391,9 @@ describe('IgxGrid - Row Selection', () => {
     }));
 
     it('Should have checkbox on each row if rowSelectable is true', (async () => {
-        const fix = TestBed.createComponent(GridWithScrollsComponent);
+        const fix = TestBed.createComponent(ScrollsComponent);
         fix.detectChanges();
-        const grid = fix.componentInstance.gridSelection5;
+        const grid = fix.componentInstance.grid;
 
         grid.rowSelectable = false;
 
@@ -480,10 +456,10 @@ describe('IgxGrid - Row Selection', () => {
     });
 
     it('Should be able to select/deselect rows programatically', fakeAsync(() => {
-        const fix = TestBed.createComponent(GridWithSelectionComponent);
+        const fix = TestBed.createComponent(SelectionComponent);
         fix.detectChanges();
 
-        const grid = fix.componentInstance.gridSelection3;
+        const grid = fix.componentInstance.grid;
         let rowsCollection = [];
         const firstRow = grid.getRowByKey('0_0');
         const secondRow = grid.getRowByKey('0_1');
@@ -532,10 +508,10 @@ describe('IgxGrid - Row Selection', () => {
     }));
 
     it('Should be able to select/deselect ALL rows programatically', fakeAsync(() => {
-        const fix = TestBed.createComponent(GridWithSelectionComponent);
+        const fix = TestBed.createComponent(SelectionComponent);
         fix.detectChanges();
 
-        const grid = fix.componentInstance.gridSelection3;
+        const grid = fix.componentInstance.grid;
         let rowsCollection = [];
         const firstRow = grid.getRowByKey('0_0');
 
@@ -696,10 +672,10 @@ describe('IgxGrid - Row Selection', () => {
     });
 
     it('Should have persistent selection through data operations - sorting', fakeAsync(() => {
-        const fix = TestBed.createComponent(GridWithSelectionComponent);
+        const fix = TestBed.createComponent(SelectionComponent);
         fix.detectChanges();
 
-        const grid = fix.componentInstance.gridSelection3;
+        const grid = fix.componentInstance.grid;
         const headerRow: HTMLElement = fix.nativeElement.querySelector('.igx-grid__thead');
         const firstRow = grid.getRowByIndex(0);
         const secondRow = grid.getRowByIndex(1);
@@ -867,10 +843,10 @@ describe('IgxGrid - Row Selection', () => {
     });
 
     it('Should be able to correctly select all rows programatically', fakeAsync(() => {
-        const fixture = TestBed.createComponent(GridWithSelectionComponent);
+        const fixture = TestBed.createComponent(SelectionComponent);
         fixture.detectChanges();
 
-        const grid = fixture.componentInstance.gridSelection3;
+        const grid = fixture.componentInstance.grid;
         const firstRow = grid.getRowByIndex(0);
         const secondRow = grid.getRowByIndex(1);
         const firstRowCheckbox: HTMLElement = firstRow.nativeElement.querySelector('.igx-checkbox__input');
@@ -893,7 +869,7 @@ describe('IgxGrid - Row Selection', () => {
         const fix = TestBed.createComponent(GridWithPrimaryKeyComponent);
         fix.detectChanges();
 
-        const grid = fix.componentInstance.gridSelection1;
+        const grid = fix.componentInstance.grid;
         grid.selectAllRows();
         fix.detectChanges();
 
@@ -981,23 +957,6 @@ describe('IgxGrid - Row Selection', () => {
 
 @Component({
     template: `
-        <igx-grid #gridSelection1 [data]="data" [primaryKey]="'ID'">
-            <igx-column field="ID"></igx-column>
-            <igx-column field="Name"></igx-column>
-            <igx-column field="JobTitle"></igx-column>
-            <igx-column field="HireDate"></igx-column>
-        </igx-grid>
-    `
-})
-export class GridWithPrimaryKeyComponent {
-    public data = data;
-
-    @ViewChild('gridSelection1', {read: IgxGridComponent})
-    public gridSelection1: IgxGridComponent;
-}
-
-@Component({
-    template: `
         <igx-grid #gridSelection2 [height]="'600px'" [data]="data" [primaryKey]="'ID'"
         [autoGenerate]="true" [rowSelectable]="true" [paging]="true" [perPage]="50">
         </igx-grid>
@@ -1041,34 +1000,6 @@ export class GridWithPagingAndSelectionComponent implements OnInit {
     }
 }
 
-@Component({
-    template: `
-        <igx-grid #gridSelection3 [data]="data" [primaryKey]="'ID'" [width]="'800px'" [height]="'600px'"
-                  [autoGenerate]="true" [rowSelectable]="true">
-        </igx-grid>
-    `
-})
-export class GridWithSelectionComponent implements OnInit {
-    public data = [];
-
-    @ViewChild('gridSelection3', {read: IgxGridComponent})
-    public gridSelection3: IgxGridComponent;
-
-    ngOnInit() {
-        const bigData = [];
-        for (let i = 0; i < 100; i++) {
-            for (let j = 0; j < 5; j++) {
-                bigData.push({
-                    ID: i.toString() + '_' + j.toString(),
-                    Column1: i * j,
-                    Column2: i * j * Math.pow(10, i),
-                    Column3: i * j * Math.pow(100, i)
-                });
-            }
-        }
-        this.data = bigData;
-    }
-}
 
 @Component({
     template: `
@@ -1090,104 +1021,9 @@ export class GridWithSelectionFilteringComponent {
     @ViewChild('gridSelection4', {read: IgxGridComponent})
     public gridSelection4: IgxGridComponent;
 
-    public data = [
-        {
-            Downloads: 254,
-            ID: 1,
-            ProductName: 'Ignite UI for JavaScript',
-            ReleaseDate: this.timeGenerator.timedelta(this.today, 'day', 15),
-            Released: false
-        },
-        {
-            Downloads: 127,
-            ID: 2,
-            ProductName: 'NetAdvantage',
-            ReleaseDate: this.timeGenerator.timedelta(this.today, 'month', -1),
-            Released: true
-        },
-        {
-            Downloads: 20,
-            ID: 3,
-            ProductName: 'Ignite UI for Angular',
-            ReleaseDate: null,
-            Released: null
-        },
-        {
-            Downloads: null,
-            ID: 4,
-            ProductName: null,
-            ReleaseDate: this.timeGenerator.timedelta(this.today, 'day', -1),
-            Released: true
-        },
-        {
-            Downloads: 100,
-            ID: 5,
-            ProductName: '',
-            ReleaseDate: undefined,
-            Released: ''
-        },
-        {
-            Downloads: 702,
-            ID: 6,
-            ProductName: 'Some other item with Script',
-            ReleaseDate: this.timeGenerator.timedelta(this.today, 'day', 1),
-            Released: null
-        },
-        {
-            Downloads: 0,
-            ID: 7,
-            ProductName: null,
-            ReleaseDate: this.timeGenerator.timedelta(this.today, 'month', 1),
-            Released: true
-        },
-        {
-            Downloads: 1000,
-            ID: 8,
-            ProductName: null,
-            ReleaseDate: this.today,
-            Released: false
-        }
-    ];
+    public data = SampleTestData.productInfoData();
 
     @ViewChild(IgxGridComponent) public grid: IgxGridComponent;
-}
-
-@Component({
-    template: `
-            <igx-grid #gridSelection3
-            [data]="data"
-            [primaryKey]="'ID'"
-            [width]="'800px'"
-            [height]="'600px'"
-            [autoGenerate]="true"
-            [rowSelectable]="true"
-        >
-        </igx-grid>
-    `
-})
-export class GridWithScrollsComponent implements OnInit {
-    public data = [];
-
-    @ViewChild(IgxGridComponent, {read: IgxGridComponent})
-    public gridSelection5: IgxGridComponent;
-
-    ngOnInit() {
-        this.data = this.getData();
-    }
-
-    public getData(rows: number = 16, cols: number = 16): any[] {
-        const bigData = [];
-        for (let i = 0; i < rows; i++) {
-            const row = {};
-            row['ID'] = i.toString();
-            for (let j = 1; j < cols; j++) {
-                row['Column ' + j] = i * j;
-            }
-
-            bigData.push(row);
-        }
-        return bigData;
-    }
 }
 
 @Component({
@@ -1208,21 +1044,9 @@ export class GridWithScrollsComponent implements OnInit {
 })
 export class GridSummaryComponent {
 
-    public data = [
-        {ProductID: 1, ProductName: 'Chai', InStock: true, UnitsInStock: 2760, OrderDate: new Date('2005-03-21')},
-        {ProductID: 2, ProductName: 'Aniseed Syrup', InStock: false, UnitsInStock: 198, OrderDate: new Date('2008-01-15')},
-        {ProductID: 3, ProductName: 'Chef Antons Cajun Seasoning', InStock: true, UnitsInStock: 52, OrderDate: new Date('2010-11-20')},
-        {ProductID: 4, ProductName: 'Grandmas Boysenberry Spread', InStock: false, UnitsInStock: 0, OrderDate: new Date('2007-10-11')},
-        {ProductID: 5, ProductName: 'Uncle Bobs Dried Pears', InStock: false, UnitsInStock: 0, OrderDate: new Date('2001-07-27')},
-        {ProductID: 6, ProductName: 'Northwoods Cranberry Sauce', InStock: true, UnitsInStock: 1098, OrderDate: new Date('1990-05-17')},
-        {ProductID: 7, ProductName: 'Queso Cabrales', InStock: false, UnitsInStock: 0, OrderDate: new Date('2005-03-03')},
-        {ProductID: 8, ProductName: 'Tofu', InStock: true, UnitsInStock: 7898, OrderDate: new Date('2017-09-09')},
-        {ProductID: 9, ProductName: 'Teatime Chocolate Biscuits', InStock: true, UnitsInStock: 6998, OrderDate: new Date('2025-12-25')},
-        {ProductID: 10, ProductName: 'Chocolate', InStock: true, UnitsInStock: 20000, OrderDate: new Date('2018-03-01')}
-    ];
+    public data = SampleTestData.foodProductData();
     @ViewChild('grid1', {read: IgxGridComponent})
     public gridSummaries: IgxGridComponent;
-
 }
 
 @Component({
@@ -1243,18 +1067,7 @@ export class GridSummaryComponent {
 })
 export class GridCancelableComponent {
 
-    public data = [
-        {ProductID: 1, ProductName: 'Chai', InStock: true, UnitsInStock: 2760, OrderDate: new Date('2005-03-21')},
-        {ProductID: 2, ProductName: 'Aniseed Syrup', InStock: false, UnitsInStock: 198, OrderDate: new Date('2008-01-15')},
-        {ProductID: 3, ProductName: 'Chef Antons Cajun Seasoning', InStock: true, UnitsInStock: 52, OrderDate: new Date('2010-11-20')},
-        {ProductID: 4, ProductName: 'Grandmas Boysenberry Spread', InStock: false, UnitsInStock: 0, OrderDate: new Date('2007-10-11')},
-        {ProductID: 5, ProductName: 'Uncle Bobs Dried Pears', InStock: false, UnitsInStock: 0, OrderDate: new Date('2001-07-27')},
-        {ProductID: 6, ProductName: 'Northwoods Cranberry Sauce', InStock: true, UnitsInStock: 1098, OrderDate: new Date('1990-05-17')},
-        {ProductID: 7, ProductName: 'Queso Cabrales', InStock: false, UnitsInStock: 0, OrderDate: new Date('2005-03-03')},
-        {ProductID: 8, ProductName: 'Tofu', InStock: true, UnitsInStock: 7898, OrderDate: new Date('2017-09-09')},
-        {ProductID: 9, ProductName: 'Teatime Chocolate Biscuits', InStock: true, UnitsInStock: 6998, OrderDate: new Date('2025-12-25')},
-        {ProductID: 10, ProductName: 'Chocolate', InStock: true, UnitsInStock: 20000, OrderDate: new Date('2018-03-01')}
-    ];
+    public data = SampleTestData.foodProductData();
     @ViewChild('gridCancelable', {read: IgxGridComponent})
     public gridCancelable: IgxGridComponent;
