+++ projects/igniteui-angular/src/lib/test-utils/helper-utils.spec.ts
@@ -1,6 +1,6 @@
 import { DebugElement } from '@angular/core';
 import { By } from '@angular/platform-browser';
-import { IgxCheckboxComponent, IgxGridComponent } from '../../public_api';
+import { IgxCheckboxComponent, IgxGridComponent, IgxGridCellComponent } from '../../public_api';
 import { wait } from '../test-utils/ui-interactions.spec';
 import { take } from 'rxjs/operators';
 import { IgxGridGroupByRowComponent } from '../grids/grid/groupby-row.component';
@@ -9,7 +9,7 @@ export class HelperUtils {
     public static getCheckboxElement(name: string, element: DebugElement, fix) {
         const checkboxElements = element.queryAll(By.css('igx-checkbox'));
         const chkElement = checkboxElements.find((el) =>
-        (el.context as IgxCheckboxComponent).placeholderLabel.nativeElement.innerText === name);
+            (el.context as IgxCheckboxComponent).placeholderLabel.nativeElement.innerText === name);
 
         return chkElement;
     }
@@ -52,46 +52,121 @@ export class HelperUtils {
         grid: IgxGridComponent,
         rowStartIndex: number,
         rowEndIndex: number,
-        colIndex?: number) => new Promise(async(resolve, reject) => {
-        const dir = rowStartIndex > rowEndIndex ? 'ArrowUp' : 'ArrowDown';
-        const row = grid.getRowByIndex(rowStartIndex);
-        const cIndx = colIndex || 0;
-        const colKey = grid.columnList.toArray()[cIndx].field;
-        let nextRow = dir === 'ArrowUp' ? grid.getRowByIndex(rowStartIndex - 1) : grid.getRowByIndex(rowStartIndex + 1);
-        const elem = row instanceof IgxGridGroupByRowComponent ?
-            row : grid.getCellByColumn(row.index, colKey);
-        if (rowStartIndex === rowEndIndex) {
-            if (!elem.focused) {
-                elem.nativeElement.focus();
-            }
-            resolve();
-            return;
-        }
-        const keyboardEvent = new KeyboardEvent('keydown', {
-            code: dir,
-            key: dir
-        });
-
-        if (dir === 'ArrowDown') {
-            elem.nativeElement.dispatchEvent(keyboardEvent);
-        } else {
-            elem.nativeElement.dispatchEvent(keyboardEvent);
-        }
-
-        if (nextRow) {
-            await wait(10);
-            HelperUtils.navigateVerticallyToIndex(grid, nextRow.index, rowEndIndex, colIndex)
-            .then(() => { resolve(); });
-        } else {
-            // else wait for chunk to load.
-            grid.verticalScrollContainer.onChunkLoad.pipe(take(1)).subscribe({
-                next: async() => {
-                    nextRow = dir === 'ArrowUp' ? grid.getRowByIndex(rowStartIndex - 1) : grid.getRowByIndex(rowStartIndex + 1);
-                    HelperUtils.navigateVerticallyToIndex(grid, nextRow.index, rowEndIndex, colIndex)
-                    .then(() => { resolve(); });
+        colIndex?: number) => new Promise(async (resolve, reject) => {
+            const dir = rowStartIndex > rowEndIndex ? 'ArrowUp' : 'ArrowDown';
+            const row = grid.getRowByIndex(rowStartIndex);
+            const cIndx = colIndex || 0;
+            const colKey = grid.columnList.toArray()[cIndx].field;
+            let nextRow = dir === 'ArrowUp' ? grid.getRowByIndex(rowStartIndex - 1) : grid.getRowByIndex(rowStartIndex + 1);
+            const elem = row instanceof IgxGridGroupByRowComponent ?
+                row : grid.getCellByColumn(row.index, colKey);
+            if (rowStartIndex === rowEndIndex) {
+                if (!elem.focused) {
+                    elem.nativeElement.focus();
                 }
+                resolve();
+                return;
+            }
+            const keyboardEvent = new KeyboardEvent('keydown', {
+                code: dir,
+                key: dir
             });
-        }
 
-    })
+            elem.nativeElement.dispatchEvent(keyboardEvent);
+
+            if (nextRow) {
+                await wait(10);
+                HelperUtils.navigateVerticallyToIndex(grid, nextRow.index, rowEndIndex, colIndex)
+                    .then(() => { resolve(); });
+            } else {
+                // else wait for chunk to load.
+                grid.verticalScrollContainer.onChunkLoad.pipe(take(1)).subscribe({
+                    next: async () => {
+                        nextRow = dir === 'ArrowUp' ? grid.getRowByIndex(rowStartIndex - 1) : grid.getRowByIndex(rowStartIndex + 1);
+                        HelperUtils.navigateVerticallyToIndex(grid, nextRow.index, rowEndIndex, colIndex)
+                            .then(() => { resolve(); });
+                    }
+                });
+            }
+
+        })
+
+    public static navigateHorizontallyToIndex = (
+        grid: IgxGridComponent,
+        cell: IgxGridCellComponent,
+        index: number) => new Promise(async (resolve) => {
+            // grid - the grid in which to navigate.
+            // cell - current cell from which the navigation will start.
+            // index - the index to which to navigate
+
+            const currIndex = cell.visibleColumnIndex;
+            const dir = currIndex < index ? 'ArrowRight' : 'ArrowLeft';
+            const nextIndex = dir === 'ArrowRight' ? currIndex + 1 : currIndex - 1;
+            const visibleColumns = grid.visibleColumns.sort((c1, c2) => c1.visibleIndex - c2.visibleIndex);
+            const nextCol = visibleColumns[nextIndex];
+            let nextCell = nextCol ? grid.getCellByColumn(0, nextCol.field) : null;
+            const keyboardEvent = new KeyboardEvent('keydown', {
+                code: dir,
+                key: dir
+            });
+            if (!cell.focused) {
+                cell.nativeElement.dispatchEvent(new Event('focus'));
+                await wait(10);
+                grid.cdr.detectChanges();
+            }
+            // if index reached return
+            if (currIndex === index) { resolve(); return; }
+            // else call arrow up/down
+            cell.nativeElement.dispatchEvent(keyboardEvent);
+
+            grid.cdr.detectChanges();
+            // if next row exists navigate next
+            if (nextCell) {
+                await wait(10);
+                grid.cdr.detectChanges();
+                HelperUtils.navigateHorizontallyToIndex(grid, nextCell, index).then(() => { resolve(); });
+            } else {
+                // else wait for chunk to load.
+                grid.parentVirtDir.onChunkLoad.pipe(take(1)).subscribe({
+                    next: () => {
+                        grid.cdr.detectChanges();
+                        nextCell = nextCol ? grid.getCellByColumn(0, nextCol.field) : null;
+                        HelperUtils.navigateHorizontallyToIndex(grid, nextCell, index).then(() => { resolve(); });
+                    }
+                });
+            }
+        })
+
+    public static expandCollapceGroupRow =
+        (fix, groupRow: IgxGridGroupByRowComponent,
+            cell: IgxGridCellComponent) => new Promise(async (resolve, reject) => {
+                expect(groupRow.focused).toBe(true);
+                expect(groupRow.nativeElement.classList.contains('igx-grid__group-row--active')).toBe(true);
+                if (cell != null) {
+                    expect(cell.selected).toBe(true);
+                }
+
+                groupRow.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'arrowleft', code: 'arrowleft', altKey: true }));
+                await wait(300);
+                fix.detectChanges();
+
+                expect(groupRow.expanded).toBe(false);
+                expect(groupRow.focused).toBe(true);
+                expect(groupRow.nativeElement.classList.contains('igx-grid__group-row--active')).toBe(true);
+                if (cell != null) {
+                    expect(cell.selected).toBe(true);
+                }
+
+                groupRow.nativeElement.dispatchEvent(new KeyboardEvent('keydown', { key: 'arrowright', code: 'arrowright', altKey: true }));
+                await wait(100);
+                fix.detectChanges();
+
+                expect(groupRow.expanded).toBe(true);
+                expect(groupRow.focused).toBe(true);
+                expect(groupRow.nativeElement.classList.contains('igx-grid__group-row--active')).toBe(true);
+                if (cell != null) {
+                    expect(cell.selected).toBe(true);
+                }
+                resolve();
+            })
 }
